requirements.txt
aio-pika==9.3.0
aioamqp==0.15.0
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiokafka==0.12.0
aioredis==2.0.1
aiormq==6.7.7
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
boto3==1.37.38
botocore==1.37.38
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.20
django-celery-beat==2.8.0
django-minio-backend==3.8.0
django-prometheus==2.3.1
django-redis==5.4.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
frozenlist==1.6.0
hiredis==3.1.0
idna==3.10
jmespath==1.0.1
kombu==5.5.3
magic-filter==1.0.12
minio==7.2.15
multidict==6.4.3
packaging==25.0
pamqp==3.2.1
pika==1.3.2
pillow==11.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
s3transfer==0.11.5
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
